Plotting error fix for isat(tis=T)
There is an error in the plotting of isat when someone selects `tis = T`. 

See the example here: 
```
library(gets)
data(Nile)
isat(Nile, tis = TRUE, plot = TRUE)
isat(Nile, tis = T, plot = TRUE)
```

Note that there is no error if someone selects `tis = TRUE` - it only happens for `tis = T`. This is because the relevant section in `print.isat` takes the information from the call element and if someone selects `tis = T` rather than `TRUE`, the information is not converted to a logical value but to a symbol. Hence the printing overall fails. 

With this fix, the value in the call is first converted to a character and then to a logical value. This should deal with the issue.