arx() output unambiguous binding
The $ operator in R auto-completes names. Hence, when the list "out" has no element called "vcov" but an element "vcov.mean" then is.null(out$vcov) still returns FALSE because R internally auto-completes the name and evaluates is.null(out$vcov.mean) instead. This defeats the purpose of this if-else statement. (Here this is not a problem because the else part uses out$vcov.mean anyways but in other settings this may cause issues.) For clarity, I have replaced the referencing using $ with [[, which is not subject to auto-complete.